requirements.txt
Flask==1.1.4
psycopg2-binary==2.8.0
mako==1.1.0
jinja2==2.10.0
markupsafe==2.0.1
gevent==20.12.0
ddtrace==2.21.5


